chore: code improvements
